pyproject.toml
[tool.poetry]
name = "invoices"
version = "0.1.0"
description = ""
authors = ["Thenuja Viknarajah <thenuja_1@outlook.com>"]
readme = "README.md"

[tool.poetry.scripts]
invoices = "invoices:app"

[tool.poetry.dependencies]
python = "^3.11"
pendulum = "^3.0.0"
pyotp = "^2.9.0"
requests = "^2.31.0"
Jinja2 = "^3.1.3"
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
typer = "^0.11.0"
python-box = "^7.1.1"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.8.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
